use gradle as project management
